refactor: 監査指摘に基づくアーキテクチャ事前補強（イベント上限拡張・フェーズガード・コマンド識別）
- GameState.FiredEventMask を ulong に拡張（ID上限解除）
- GameFlowController.BeginTurn のイベント待機フェーズ制御を修正
- CommandResolverSO に CommandDataSO 引数のオーバーロードを追加
- 47/47 EditMode テスト全件通過確認

+++ Game/Assets/Core/Data/GameState.cs
@@ -17,7 +17,8 @@ namespace Game.Core
 
         /// <summary>
         /// 発火済みイベントのビットマスク。イベント ID n が発火済みのとき、ビット n が立つ。
+        /// ulong（64ビット）とすることで、イベント ID 0〜63 まで扱える。
         /// </summary>
-        public int FiredEventMask { get; init; }
+        public ulong FiredEventMask { get; init; }
     }
 }
